feat: enhance modularity by relocating core packages
