+++ cmd/worker/main.go
@@ -5,8 +5,8 @@ package main
 
 import (
 	"github.com/labstack/echo/v4"
-	"github.com/sonr-io/motr/pkg/database"
-	"github.com/sonr-io/motr/pkg/session"
+	"github.com/sonr-io/motr/middleware/database"
+	"github.com/sonr-io/motr/middleware/session"
 	"github.com/sonr-io/motr/routes"
 	"github.com/syumai/workers"
 	_ "github.com/syumai/workers/cloudflare/d1"
